// src/generated/model/classResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ClassResponseStatus } from './classResponseStatus.ts';

export interface ClassResponse {
  id?: string;
  programId?: string;
  programLevelId?: string;
  batchId?: string;
  code?: string;
  name?: string;
  capacity?: number;
  status?: ClassResponseStatus;
  remarks?: string;
  createdAt?: string;
  updatedAt?: string;
}