// src/generated/model/programLevelSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ProgramLevelSummaryStatus } from './programLevelSummaryStatus.ts';

export interface ProgramLevelSummary {
  id?: string;
  code?: string;
  name?: string;
  status?: ProgramLevelSummaryStatus;
  displayOrder?: number;
}